feat(zsh): add aliases: `prname`, `ccr`, `silent`, `si`
- Remove alias `rename=perl-rename`
- Add alias `prname=perl-rename`
- Add alias `ccr=cookiecutter`
- Add function `silent`

Especially since the `rename=perl-rename` alias was only active when
`perl-rename` was available, the `rename` alias was a bad idea. It meant
that I had to check `which rename` every time I wanted to use the
command. Having a distinct alias for perl-rename(1) means I can use it
and also have the system tell me when it's unavailable, instead of
silently falling back to rename(1).

`silent` and its alias are useful because I frequently
want to spawn `xdg-open` etc without actually using it.